Rearranged panels with day-night graphic
Rearranged pannels so that they show a day/night icon on the left with forecast info on the right. Added dictionaries and icons to support forecast icons later on. Revised readme to include icon attributions. forecastImageAPI includes a list of forcasts and the image inks for the NOAA API, but is unused for now.