Added updat forecast button
In order to get the update forecast button working in the navbar, I restructured the code, so now the API requests are promises. This removes the previous 'callback hell' chaining that used to make up the structure. Error handling still needs to be added/tessted. However, since the functions no longer call eachother in succession, it should be easier to update parts of the data dynamically.